perf: Save paddles score in a byte instead of int32

+++ paddle.go
@@ -6,7 +6,7 @@ type Paddle struct {
 	rl.Rectangle
 
 	Speed float
-	Score int
+	Score byte
 }
 
 func NewPaddle() *Paddle {
